make builder,unstable_discord_api combination work
