+++ src/builder/create_message.rs
@@ -4,7 +4,7 @@ use super::CreateAllowedMentions;
 use super::CreateEmbed;
 #[cfg(feature = "unstable_discord_api")]
 use crate::builder::CreateComponents;
-use crate::http::AttachmentType;
+use crate::model::channel::AttachmentType;
 use crate::internal::prelude::*;
 use crate::json::{self, to_value};
 use crate::model::channel::{MessageReference, ReactionType};
