+++ src/builder/edit_message.rs
@@ -3,9 +3,9 @@ use std::collections::HashMap;
 use super::CreateEmbed;
 #[cfg(feature = "unstable_discord_api")]
 use crate::builder::CreateComponents;
-use crate::http::AttachmentType;
 use crate::internal::prelude::*;
 use crate::json::{self, from_number};
+use crate::model::channel::AttachmentType;
 
 /// A builder to specify the fields to edit in an existing message.
 ///
